refactor: reorganized components structure
